backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.17.0
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.25.2
langdetect==1.0.9
google-cloud-translate==2.0.1
